Factored out logic from controller into services
